FileAPI/BlobURL/Blob-File-URL-Test1.html

<!DOCTYPE html>
<html>
<body>
<head>
<meta charset=utf-8>
<title>Blob and File reference URL Test(1)</title>
<link rel=help href="http://www.w3.org/TR/FileAPI/#convenienceAPI">
<link rel=author title="Breezewish" href="mailto:me@breeswish.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <form name="upload">
        <input type="file" id="fileChooser">
    </form>

    <div>
        <p>Test steps:</p>
        <ol>
            <li>Download the <a href="support/file_test1.js">file</a>.</li>
            <li>Select the file in the file inputbox to run the test.</li>
        </ol>
    </div>

    <div id="log"></div>

    <script>

        var fileChooser = document.querySelector('#fileChooser');

        setup({explicit_done: true});
        setup({explicit_timeout: true});

        //Run the test when user selects a file

        on_event(fileChooser, 'change', function() {


            test(function() {
                assert_true("URL" in window, "window should have a 'URL' interface.");
                assert_true("createObjectURL" in window.URL, "window.URL should have a 'createObjectURL' function.");
                assert_true("revokeObjectURL" in window.URL, "window.URL should have a 'revokeObjectURL' function.");
                assert_true(URL.createObjectURL instanceof Function, "URL.createObjectURL should be a function.");
                assert_true(URL.revokeObjectURL instanceof Function, "URL.revokeObjectURL should be a function.");
            }, "Check whether Blob/File URL interface is implemented");




            var testCount = 10000;

            test(function() {
                
                var list = [], file = fileChooser.files[0];
                

                for (var i = 0; i <= testCount; i++) {
                    list.push(window.URL.createObjectURL(file));
                }

                list.sort();

                for (var i = 0; i < testCount; i++) {
                    assert_not_equals(list[i], list[i+1], 'generated Blob URL should be unique');
                }

            }, 'Check whether generated Blob/File URL is unique (Notice: only generate for ' + testCount + ' times)');







            var t1 = async_test('Check whether Blob/File URL could be used in XHR requests and could get expected data');
            t1.step(function() {


                var url = URL.createObjectURL(fileChooser.files[0]);
                var expected_file_content = "var test_result = 'OK';";

                var xhr = new XMLHttpRequest();
                xhr.open('GET', url, true);
                xhr.onreadystatechange = function(e) {

                    if (this.readyState == 4) {

                        t1.step_func_done(function() {

                            assert_equals(this.status, 200, 'status code should be 200');
                            assert_equals(this.responseText, expected_file_content);

                            t1.done();

                        }, this)();

                    }
                }
                xhr.send();

            });








            var t2 = async_test('Check whether Blob/File URL could be used in tags src like <script>');
            t2.step(function() {


                var url = URL.createObjectURL(fileChooser.files[0]);
                var expected_run_result = "OK";
                
                //expected file content:
                //   var test_result = 'OK';

                var e = document.createElement('script');
                e.setAttribute('type', 'text/javascript');
                e.setAttribute('src', url);

                e.onload = t2.step_func_done(function() {

                    assert_equals(test_result, expected_run_result);
                    t2.done();

                }, e);

                document.body.appendChild(e);

            });








            var t3 = async_test('Check whether revokeObjectURL works well');
            t3.step(function() {


                var url = URL.createObjectURL(fileChooser.files[0]);
                URL.revokeObjectURL(url);

                var xhr = new XMLHttpRequest();
                xhr.open('GET', url, true);
                xhr.onreadystatechange = function(e) {

                    if (this.readyState == 4) {

                        t3.step_func_done(function() {

                            assert_equals(this.status, 500, 'status code should be 500 if Blob URI is revoked.');
                            t3.done();

                        }, this)();

                    }
                }
                xhr.send();

            });





            done();

        });
        

    </script>
</body>
</html>
